requirements.txt
pyaes==1.6.1 ; python_version >= "3.10" and python_version < "4.0" \
    --hash=sha256:02c1b1405c38d3c370b085fb952dd8bea3fadcee6411ad99f312cc129c536d8f
pyasn1==0.5.1 ; python_version >= "3.10" and python_version < "4" \
    --hash=sha256:4439847c58d40b1d0a573d07e3856e95333f1976294494c325775aeca506eb58 \
    --hash=sha256:6d391a96e59b23130a5cfa74d6fd7f388dbbe26cc8f1edf39fdddf08d9d6676c
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0" \
    --hash=sha256:e324ee90a023d808f1959c46bcbc04446a10ced277783dc6ee09987c37ec10ca \
    --hash=sha256:f7b63ef50f1b690dddf550d03497b66d609393b40b564ed0d674909a68ebf16a
rsa==4.9 ; python_version >= "3.10" and python_version < "4" \
    --hash=sha256:90260d9058e514786967344d0ef75fa8727eed8a7d2e43ce9f4bcf1b536174f7 \
    --hash=sha256:e38464a49c6c85d7f1351b0126661487a7e0a14a50f1675ec50eb34d4f20ef21
telethon==1.34.0 ; python_version >= "3.10" and python_version < "4.0" \
    --hash=sha256:55290809a30081fa0bb5052abb7547cbb25d7fbca94f32f13c147504d521804f
